# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gadget X Repairs API — lead capture, inventory browse, admin leads view.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - in: query
          name: category
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryItem" }
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a single inventory item by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItem" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/leads:
    get:
      operationId: getAdminLeads
      tags: [admin]
      summary: Get all leads grouped by type (password-gated)
      parameters:
        - in: header
          name: X-Admin-Password
          required: true
          schema: { type: string }
      responses:
        "200":
          description: All leads grouped
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminLeads" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/leads/{leadType}/{id}/status:
    patch:
      operationId: updateLeadStatus
      tags: [admin]
      summary: Update a lead status (password-gated)
      parameters:
        - in: header
          name: X-Admin-Password
          required: true
          schema: { type: string }
        - in: path
          name: leadType
          required: true
          schema:
            type: string
            enum:
              - repair-quote
              - sell-phone
              - appointment
              - contact
              - reservation
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadStatusUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmitResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/messaging/config:
    get:
      operationId: getMessagingConfig
      tags: [admin]
      summary: Get whether email and SMS sending are configured
      parameters:
        - in: header
          name: X-Admin-Password
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Messaging config
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessagingConfig" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/leads/{leadType}/{id}/activity:
    get:
      operationId: getLeadActivity
      tags: [admin]
      summary: Get communication activity for a single lead
      parameters:
        - in: header
          name: X-Admin-Password
          required: true
          schema: { type: string }
        - in: path
          name: leadType
          required: true
          schema:
            type: string
            enum:
              - repair-quote
              - sell-phone
              - appointment
              - contact
              - reservation
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeadActivityResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/leads/{leadType}/{id}/email:
    post:
      operationId: sendLeadEmail
      tags: [admin]
      summary: Send an email reply to a lead
      parameters:
        - in: header
          name: X-Admin-Password
          required: true
          schema: { type: string }
        - in: path
          name: leadType
          required: true
          schema:
            type: string
            enum:
              - repair-quote
              - sell-phone
              - appointment
              - contact
              - reservation
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendEmailInput" }
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendResult" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          description: Rate limited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/leads/{leadType}/{id}/sms:
    post:
      operationId: sendLeadSms
      tags: [admin]
      summary: Send an SMS reply to a lead
      parameters:
        - in: header
          name: X-Admin-Password
          required: true
          schema: { type: string }
        - in: path
          name: leadType
          required: true
          schema:
            type: string
            enum:
              - repair-quote
              - sell-phone
              - appointment
              - contact
              - reservation
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendSmsInput" }
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendResult" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          description: Rate limited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    SubmitResponse:
      type: object
      properties:
        ok: { type: boolean }
        id: { type: integer }
      required: [ok]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        details: { type: object, additionalProperties: true }
      required: [error]
    RepairQuoteInput:
      type: object
      required: [name, phone, deviceType, brand, model, problem]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        phone: { type: string, minLength: 7, maxLength: 40 }
        email: { type: string, maxLength: 200 }
        deviceType: { type: string, minLength: 1, maxLength: 80 }
        brand: { type: string, minLength: 1, maxLength: 80 }
        model: { type: string, minLength: 1, maxLength: 120 }
        problem: { type: string, minLength: 1, maxLength: 4000 }
        preferredContact:
          type: string
          enum: [call, text, whatsapp, email]
        urgency:
          type: string
          enum: [asap, today, this_week, flexible]
        notes: { type: string, maxLength: 2000 }
        photoUrl: { type: string, maxLength: 500 }
    SellPhoneInput:
      type: object
      required: [name, phone, brand, model]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        phone: { type: string, minLength: 7, maxLength: 40 }
        brand: { type: string, minLength: 1, maxLength: 80 }
        model: { type: string, minLength: 1, maxLength: 120 }
        storage: { type: string, maxLength: 40 }
        carrier: { type: string, maxLength: 80 }
        lockedStatus:
          type: string
          enum: [locked, unlocked]
        condition:
          type: string
          enum: [mint, good, fair, broken]
        batteryHealth: { type: integer, minimum: 0, maximum: 100 }
        damageNotes: { type: string, maxLength: 2000 }
        expectedPrice: { type: string, maxLength: 20 }
        photoUrl: { type: string, maxLength: 500 }
    AppointmentInput:
      type: object
      required: [name, phone, serviceType, preferredDatetime]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        phone: { type: string, minLength: 7, maxLength: 40 }
        serviceType:
          type: string
          enum:
            - screen-repair
            - battery-replacement
            - hdmi-repair
            - laptop-diagnostic
            - motherboard-repair
            - phone-unlocking
            - prepaid-activation
        preferredDatetime: { type: string, minLength: 1, maxLength: 80 }
        notes: { type: string, maxLength: 2000 }
    ContactInput:
      type: object
      required: [name, contact, message]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        contact: { type: string, minLength: 3, maxLength: 200 }
        message: { type: string, minLength: 1, maxLength: 4000 }
    ReservationInput:
      type: object
      required: [name, phone, itemId, itemLabel]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        phone: { type: string, minLength: 7, maxLength: 40 }
        itemId: { type: string, minLength: 1, maxLength: 120 }
        itemLabel: { type: string, minLength: 1, maxLength: 200 }
        notes: { type: string, maxLength: 1000 }
    InventoryItem:
      type: object
      required: [id, category, brand, model, price]
      properties:
        id: { type: string }
        category: { type: string }
        brand: { type: string }
        model: { type: string }
        storage: { type: string }
        color: { type: string }
        condition: { type: string }
        carrier: { type: string }
        price: { type: string }
        warranty: { type: string }
        availability: { type: string }
        imageUrl: { type: string }
        description: { type: string }
    LeadBase:
      type: object
      properties:
        id: { type: integer }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, status, createdAt, updatedAt]
    RepairQuoteRecord:
      allOf:
        - $ref: "#/components/schemas/LeadBase"
        - $ref: "#/components/schemas/RepairQuoteInput"
    SellPhoneRecord:
      allOf:
        - $ref: "#/components/schemas/LeadBase"
        - $ref: "#/components/schemas/SellPhoneInput"
    AppointmentRecord:
      allOf:
        - $ref: "#/components/schemas/LeadBase"
        - $ref: "#/components/schemas/AppointmentInput"
    ContactRecord:
      allOf:
        - $ref: "#/components/schemas/LeadBase"
        - $ref: "#/components/schemas/ContactInput"
    ReservationRecord:
      allOf:
        - $ref: "#/components/schemas/LeadBase"
        - $ref: "#/components/schemas/ReservationInput"
    LeadStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, in_progress, done, archived]
    MessagingConfig:
      type: object
      required: [emailEnabled, smsEnabled, mailFrom, smsFrom]
      properties:
        emailEnabled: { type: boolean }
        smsEnabled: { type: boolean }
        mailFrom: { type: string }
        smsFrom: { type: string }
    LeadCommunication:
      type: object
      required:
        [
          id,
          leadType,
          leadId,
          channel,
          direction,
          body,
          recipient,
          status,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        leadType:
          type: string
          enum:
            - repair-quote
            - sell-phone
            - appointment
            - contact
            - reservation
        leadId: { type: string }
        channel:
          type: string
          enum: [email, sms]
        direction:
          type: string
          enum: [outbound, inbound]
        subject: { type: string, nullable: true }
        body: { type: string }
        recipient: { type: string }
        providerMessageId: { type: string, nullable: true }
        status: { type: string }
        error: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
    LeadActivityResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/LeadCommunication" }
    SendEmailInput:
      type: object
      required: [to, subject]
      properties:
        to: { type: string, minLength: 3, maxLength: 200 }
        subject: { type: string, minLength: 1, maxLength: 200 }
        html: { type: string, maxLength: 20000 }
        text: { type: string, maxLength: 20000 }
    SendSmsInput:
      type: object
      required: [to, body]
      properties:
        to: { type: string, minLength: 7, maxLength: 40 }
        body: { type: string, minLength: 1, maxLength: 1600 }
    SendResult:
      type: object
      required: [ok, status]
      properties:
        ok: { type: boolean }
        id: { type: integer }
        providerMessageId: { type: string, nullable: true }
        status: { type: string, enum: [queued, sent, failed] }
        error: { type: string, nullable: true }
...
    AdminLeads:
      type: object
      required:
        [
          repairQuotes,
          sellPhoneSubmissions,
          appointments,
          contactMessages,
          itemReservations,
        ]
      properties:
        repairQuotes:
          type: array
          items: { $ref: "#/components/schemas/RepairQuoteRecord" }
        sellPhoneSubmissions:
          type: array
          items: { $ref: "#/components/schemas/SellPhoneRecord" }
        appointments:
          type: array
          items: { $ref: "#/components/schemas/AppointmentRecord" }
        contactMessages:
          type: array
          items: { $ref: "#/components/schemas/ContactRecord" }
        itemReservations:
          type: array
          items: { $ref: "#/components/schemas/ReservationRecord" }